Rename `duke.internals` to `duke.internal`
