+++ src/main/java/duke/internal/Storage.java
@@ -1,4 +1,4 @@
-package duke.internals;
+package duke.internal;
 
 import duke.task.Task;
 import duke.task.TaskList;
